Reorganised Data
Added New Folders and Reorganised data for readability
