+++ IntraDayTradingSystem/CreatingStockUniverse/scraped_data.csv
@@ -1,0 +1,137 @@
+IB Symbol,Product Description,Symbol,Currency
+ACC,ACC LIMITED,ACC,INR
+ADANIENT,ADANI ENTERPRISES LTD,ADANIENT,INR
+ADANIPORT,ADANI PORTS AND SPECIAL ECON,ADANIPORTS,INR
+AMARAJABA,AMARA RAJA BATTERIES LTD,AMARAJABAT,INR
+AMBUJACEM,AMBUJA CEMENTS LIMITED,AMBUJACEM,INR
+APOLLOHOS,APOLLO HOSPITALS ENTERPRISE,APOLLOHOSP,INR
+APOLLOTYR,APOLLO TYRES LIMITED,APOLLOTYRE,INR
+ASHOKLEY,ASHOK LEYLAND LIMITED,ASHOKLEY,INR
+ASIANPAIN,ASIAN PAINTS LTD,ASIANPAINT,INR
+AUROPHARM,AUROBINDO PHARMA LTD,AUROPHARMA,INR
+AXISBANK,AXIS BANK LTD,AXISBANK,INR
+BAJAJ-AUT,BAJAJ AUTO LIMITED,BAJAJ-AUTO,INR
+BAJAJFINS,BAJAJ FINSERV LTD,BAJAJFINSV,INR
+BAJFINANC,BAJAJ FINANCE LTD,BAJFINANCE,INR
+BALKRISIN,BALKRISHNA INDUSTRIES LTD,BALKRISIND,INR
+BANDHANBN,BANDHAN BANK LTD,BANDHANBNK,INR
+BANKBAROD,BANK OF BARODA,BANKBARODA,INR
+BATAINDIA,BATA INDIA LTD,BATAINDIA,INR
+BEL,BHARAT ELECTRONICS LTD,BEL,INR
+BERGEPAIN,BERGER PAINTS INDIA LTD,BERGEPAINT,INR
+BHARATFOR,BHARAT FORGE LIMITED,BHARATFORG,INR
+BHARTIART,BHARTI AIRTEL LIMITED,BHARTIARTL,INR
+BHEL,BHARAT HEAVY ELECTRICALS,BHEL,INR
+BIOCON,BIOCON LTD,BIOCON,INR
+BOSCHLTD,BOSCH LIMITED,BOSCHLTD,INR
+BPCL,BHARAT PETROLEUM CORP LTD,BPCL,INR
+BRITANNIA,BRITANNIA INDUSTRIES LTD,BRITANNIA,INR
+CADILAHC,CADILA HEALTHCARE LTD,CADILAHC,INR
+CANBK,CANARA BANK,CANBK,INR
+CHOLAFIN,CHOLAMANDALAM INVESTMENT AND,CHOLAFIN,INR
+CIPLA,CIPLA LTD,CIPLA,INR
+COALINDIA,COAL INDIA LTD,COALINDIA,INR
+COFORGE,COFORGE LTD,COFORGE,INR
+COLPAL,COLGATE PALMOLIVE (INDIA),COLPAL,INR
+CONCOR,CONTAINER CORP OF INDIA LTD,CONCOR,INR
+CUMMINSIN,CUMMINS INDIA LTD,CUMMINSIND,INR
+DABUR,DABUR INDIA LIMITED,DABUR,INR
+DIVISLAB,DIVI'S LABORATORIES LTD,DIVISLAB,INR
+DLF,DLF LIMITED,DLF,INR
+DRREDDY,DR. REDDY'S LABORATORIES,DRREDDY,INR
+EICHERMOT,EICHER MOTORS LTD,EICHERMOT,INR
+ESCORTS,ESCORTS LTD,ESCORTS,INR
+EXIDEIND,EXIDE INDUSTRIES LTD,EXIDEIND,INR
+FEDERALBN,FEDERAL BANK LTD,FEDERALBNK,INR
+GAIL,GAIL INDIA LTD,GAIL,INR
+GLENMARK,GLENMARK PHARMACEUTICALS LTD,GLENMARK,INR
+GMRINFRA,GMR INFRASTRUCTURE LTD,GMRINFRA,INR
+GODREJCP,GODREJ CONSUMER PRODUCTS LTD,GODREJCP,INR
+GODREJPRO,GODREJ PROPERTIES LTD,GODREJPROP,INR
+GRASIM,GRASIM INDUSTRIES LTD,GRASIM,INR
+HAVELLS,HAVELLS INDIA LTD,HAVELLS,INR
+HCLTECH,HCL TECHNOLOGIES LTD,HCLTECH,INR
+HDFC,HOUSING DEVELOPMENT FINANCE,HDFC,INR
+HDFCBANK,HDFC BANK LIMITED,HDFCBANK,INR
+HDFCLIFE,HDFC LIFE INSURANCE CO LTD,HDFCLIFE,INR
+HEROMOTOC,HERO MOTOCORP LIMITED,HEROMOTOCO,INR
+HINDALCO,HINDALCO INDUSTRIES LIMITED,HINDALCO,INR
+HINDPETRO,HINDUSTAN PETROLEUM CORP,HINDPETRO,INR
+HINDUNILV,HINDUSTAN UNILEVER LTD,HINDUNILVR,INR
+IBULHSGFI,INDIABULLS HOUSING FINANCE L,IBULHSGFIN,INR
+ICICIBANK,ICICI BANK LTD,ICICIBANK,INR
+ICICIPRUL,ICICI PRUDENTIAL LIFE INSURA,ICICIPRULI,INR
+IDEA,VODAFONE IDEA LTD,IDEA,INR
+IDFCFIRST,IDFC FIRST BANK LTD,IDFCFIRSTB,INR
+IGL,INDRAPRASTHA GAS LTD,IGL,INR
+INDIGO,INTERGLOBE AVIATION LTD,INDIGO,INR
+INDUSINDB,INDUSIND BANK LIMITED,INDUSINDBK,INR
+INFRATEL,BHARTI INFRATEL LTD,INFRATEL,INR
+INFY,INFOSYS LTD,INFY,INR
+IOC,INDIAN OIL CORPORATION LTD,IOC,INR
+ITC,ITC LTD,ITC,INR
+JINDALSTE,JINDAL STEEL & POWER LTD,JINDALSTEL,INR
+JSWSTEEL,JSW STEEL LTD,JSWSTEEL,INR
+JUBLFOOD,JUBILANT FOODWORKS LTD,JUBLFOOD,INR
+KOTAKBANK,KOTAK MAHINDRA BANK LTD,KOTAKBANK,INR
+LICHSGFIN,LIC HOUSING FINANCE LTD,LICHSGFIN,INR
+LT,LARSEN & TOUBRO LIMITED,LT,INR
+LTFH,L&T FINANCE HOLDINGS LTD,L&TFH,INR
+LUPIN,LUPIN LTD,LUPIN,INR
+MANAPPURA,MANAPPURAM FINANCE LTD,MANAPPURAM,INR
+MARICO,MARICO LIMITED,MARICO,INR
+MARUTI,MARUTI SUZUKI INDIA LTD,MARUTI,INR
+MCDOWELL-,UNITED SPIRITS LTD,MCDOWELL-N,INR
+MFSL,MAX FINANCIAL SERVICES LTD,MFSL,INR
+MGL,MAHANAGAR GAS LTD,MGL,INR
+MINDTREE,MINDTREE LIMITED,MINDTREE,INR
+MM,MAHINDRA & MAHINDRA LIMITED,M&M,INR
+MMFIN,MAHINDRA & MAHINDRA FIN SECS,M&MFIN,INR
+MOTHERSUM,MOTHERSON SUMI SYSTEMS LTD,MOTHERSUMI,INR
+MRF,MRF LIMITED,MRF,INR
+MUTHOOTFI,MUTHOOT FINANCE LTD,MUTHOOTFIN,INR
+NATIONALU,NATIONAL ALUMINIUM CO LTD,NATIONALUM,INR
+NAUKRI,INFO EDGE INDIA LTD,NAUKRI,INR
+NESTLEIND,NESTLE INDIA LIMITED,NESTLEIND,INR
+NMDC,NMDC LIMITED,NMDC,INR
+NTPC,NTPC LIMITED,NTPC,INR
+ONGC,OIL & NATURAL GAS CORP LTD,ONGC,INR
+PAGEIND,PAGE INDUSTRIES LTD,PAGEIND,INR
+PEL,PIRAMAL ENTERPRISES LTD,PEL,INR
+PETRONET,PETRONET LNG LTD,PETRONET,INR
+PFC,POWER FINANCE CORPORATION,PFC,INR
+PIDILITIN,PIDILITE INDUSTRIES LTD,PIDILITIND,INR
+PNB,PUNJAB NATIONAL BANK,PNB,INR
+POWERGRID,POWER GRID CORP. OF INDIA LTD.,POWERGRID,INR
+PVR,PVR LTD,PVR,INR
+RAMCOCEM,RAMCO CEMENTS LTD/THE,RAMCOCEM,INR
+RBLBANK,RBL BANK LTD,RBLBANK,INR
+RECLTD,REC LTD,RECLTD,INR
+RELIANCE,RELIANCE INDUSTRIES LIMITED,RELIANCE,INR
+SAIL,STEEL AUTHORITY OF INDIA,SAIL,INR
+SBILIFE,SBI LIFE INSURANCE CO LTD,SBILIFE,INR
+SBIN,STATE BANK OF INDIA,SBIN,INR
+SHREECEM,SHREE CEMENT LIMITED,SHREECEM,INR
+SIEMENS,SIEMENS LTD,SIEMENS,INR
+SRF,SRF LIMITED,SRF,INR
+SRTRANSFI,SHRIRAM TRANSPORT FINANCE,SRTRANSFIN,INR
+SUNPHARMA,SUN PHARMACEUTICAL INDUS,SUNPHARMA,INR
+SUNTV,SUN TV NETWORK LIMITED,SUNTV,INR
+TATACHEM,TATA CHEMICALS LIMITED,TATACHEM,INR
+TATACONSU,TATA CONSUMER PRODUCTS LTD,TATACONSUM,INR
+TATAMOTOR,TATA MOTORS LTD,TATAMOTORS,INR
+TATAPOWER,TATA POWER CO LTD,TATAPOWER,INR
+TATASTEEL,TATA STEEL LTD,TATASTEEL,INR
+TCS,TATA CONSULTANCY SVCS LTD,TCS,INR
+TECHM,TECH MAHINDRA LTD,TECHM,INR
+TITAN,TITAN CO LTD,TITAN,INR
+TORNTPHAR,TORRENT PHARMACEUTICALS LTD,TORNTPHARM,INR
+TORNTPOWE,TORRENT POWER LTD,TORNTPOWER,INR
+TVSMOTOR,TVS MOTOR CO LTD,TVSMOTOR,INR
+UBL,UNITED BREWERIES LTD,UBL,INR
+ULTRACEMC,ULTRATECH CEMENT LTD,ULTRACEMCO,INR
+UPL,UPL LTD,UPL,INR
+VEDL,VEDANTA LTD,VEDL,INR
+VOLTAS,VOLTAS LIMITED,VOLTAS,INR
+WIPRO,WIPRO LTD,WIPRO,INR
+ZEEL,ZEE ENTERTAINMENT ENTERPRISE,ZEEL,INR
