+++ IntraDayTradingSystem/CreatingStockUniverse/stock_universe_web_scraping.csv
@@ -1,0 +1,137 @@
+IB Symbol,Product Description,Symbol,Currency,Contract ID
+ACC,ACC LIMITED,ACC,INR,44652144
+ADANIENT,ADANI ENTERPRISES LTD,ADANIENT,INR,56986798
+ADANIPORT,ADANI PORTS AND SPECIAL ECON,ADANIPORTS,INR,79403918
+AMARAJABA,AMARA RAJA BATTERIES LTD,AMARAJABAT,INR,113744943
+AMBUJACEM,AMBUJA CEMENTS LIMITED,AMBUJACEM,INR,44652129
+APOLLOHOS,APOLLO HOSPITALS ENTERPRISE,APOLLOHOSP,INR,78607402
+APOLLOTYR,APOLLO TYRES LIMITED,APOLLOTYRE,INR,56988336
+ASHOKLEY,ASHOK LEYLAND LIMITED,ASHOKLEY,INR,56988221
+ASIANPAIN,ASIAN PAINTS LTD,ASIANPAINT,INR,132214022
+AUROPHARM,AUROBINDO PHARMA LTD,AUROPHARMA,INR,83937139
+AXISBANK,AXIS BANK LTD,AXISBANK,INR,162089869
+BAJAJ-AUT,BAJAJ AUTO LIMITED,BAJAJ-AUTO,INR,56987980
+BAJAJFINS,BAJAJ FINSERV LTD,BAJAJFINSV,INR,56987991
+BAJFINANC,BAJAJ FINANCE LTD,BAJFINANCE,INR,247227232
+BALKRISIN,BALKRISHNA INDUSTRIES LTD,BALKRISIND,INR,81899870
+BANDHANBN,BANDHAN BANK LTD,BANDHANBNK,INR,312182039
+BANKBAROD,BANK OF BARODA,BANKBARODA,INR,180222138
+BATAINDIA,BATA INDIA LTD,BATAINDIA,INR,208706811
+BEL,BHARAT ELECTRONICS LTD,BEL,INR,269577540
+BERGEPAIN,BERGER PAINTS INDIA LTD,BERGEPAINT,INR,179252695
+BHARATFOR,BHARAT FORGE LIMITED,BHARATFORG,INR,56987929
+BHARTIART,BHARTI AIRTEL LIMITED,BHARTIARTL,INR,64769335
+BHEL,BHARAT HEAVY ELECTRICALS,BHEL,INR,95441327
+BIOCON,BIOCON LTD,BIOCON,INR,56987974
+BOSCHLTD,BOSCH LIMITED,BOSCHLTD,INR,56988066
+BPCL,BHARAT PETROLEUM CORP LTD,BPCL,INR,44652160
+BRITANNIA,BRITANNIA INDUSTRIES LTD,BRITANNIA,INR,343576985
+CADILAHC,CADILA HEALTHCARE LTD,CADILAHC,INR,208509801
+CANBK,CANARA BANK,CANBK,INR,56985302
+CHOLAFIN,CHOLAMANDALAM INVESTMENT AND,CHOLAFIN,INR,370754270
+CIPLA,CIPLA LTD,CIPLA,INR,44652139
+COALINDIA,COAL INDIA LTD,COALINDIA,INR,80626626
+COFORGE,COFORGE LTD,COFORGE,INR,56984634
+COLPAL,COLGATE PALMOLIVE (INDIA),COLPAL,INR,56985419
+CONCOR,CONTAINER CORP OF INDIA LTD,CONCOR,INR,322790806
+CUMMINSIN,CUMMINS INDIA LTD,CUMMINSIND,INR,56985045
+DABUR,DABUR INDIA LIMITED,DABUR,INR,44652199
+DIVISLAB,DIVI'S LABORATORIES LTD,DIVISLAB,INR,56984084
+DLF,DLF LIMITED,DLF,INR,49982064
+DRREDDY,DR. REDDY'S LABORATORIES,DRREDDY,INR,44652180
+EICHERMOT,EICHER MOTORS LTD,EICHERMOT,INR,440773528
+ESCORTS,ESCORTS LTD,ESCORTS,INR,56984411
+EXIDEIND,EXIDE INDUSTRIES LTD,EXIDEIND,INR,56985401
+FEDERALBN,FEDERAL BANK LTD,FEDERALBNK,INR,136316474
+GAIL,GAIL INDIA LTD,GAIL,INR,44652084
+GLENMARK,GLENMARK PHARMACEUTICALS LTD,GLENMARK,INR,56986291
+GMRINFRA,GMR INFRASTRUCTURE LTD,GMRINFRA,INR,69237773
+GODREJCP,GODREJ CONSUMER PRODUCTS LTD,GODREJCP,INR,56986171
+GODREJPRO,GODREJ PROPERTIES LTD,GODREJPROP,INR,138950253
+GRASIM,GRASIM INDUSTRIES LTD,GRASIM,INR,250071842
+HAVELLS,HAVELLS INDIA LTD,HAVELLS,INR,164667099
+HCLTECH,HCL TECHNOLOGIES LTD,HCLTECH,INR,44652094
+HDFC,HOUSING DEVELOPMENT FINANCE,HDFC,INR,77975259
+HDFCBANK,HDFC BANK LIMITED,HDFCBANK,INR,382750910
+HDFCLIFE,HDFC LIFE INSURANCE CO LTD,HDFCLIFE,INR,296455120
+HEROMOTOC,HERO MOTOCORP LIMITED,HEROMOTOCO,INR,44652149
+HINDALCO,HINDALCO INDUSTRIES LIMITED,HINDALCO,INR,44652119
+HINDPETRO,HINDUSTAN PETROLEUM CORP,HINDPETRO,INR,44652194
+HINDUNILV,HINDUSTAN UNILEVER LTD,HINDUNILVR,INR,44652060
+IBULHSGFI,INDIABULLS HOUSING FINANCE L,IBULHSGFIN,INR,125585152
+ICICIBANK,ICICI BANK LTD,ICICIBANK,INR,175596029
+ICICIPRUL,ICICI PRUDENTIAL LIFE INSURA,ICICIPRULI,INR,249453974
+IDEA,VODAFONE IDEA LTD,IDEA,INR,50486121
+IDFCFIRST,IDFC FIRST BANK LTD,IDFCFIRSTB,INR,209462790
+IGL,INDRAPRASTHA GAS LTD,IGL,INR,294350643
+INDIGO,INTERGLOBE AVIATION LTD,INDIGO,INR,212126639
+INDUSINDB,INDUSIND BANK LIMITED,INDUSINDBK,INR,56985651
+INFRATEL,BHARTI INFRATEL LTD,INFRATEL,INR,119707173
+INFY,INFOSYS LTD,INFY,INR,44652017
+IOC,INDIAN OIL CORPORATION LTD,IOC,INR,56985747
+ITC,ITC LTD,ITC,INR,44652040
+JINDALSTE,JINDAL STEEL & POWER LTD,JINDALSTEL,INR,56984268
+JSWSTEEL,JSW STEEL LTD,JSWSTEEL,INR,261024709
+JUBLFOOD,JUBILANT FOODWORKS LTD,JUBLFOOD,INR,72513009
+KOTAKBANK,KOTAK MAHINDRA BANK LTD,KOTAKBANK,INR,78743661
+LICHSGFIN,LIC HOUSING FINANCE LTD,LICHSGFIN,INR,82450876
+LT,LARSEN & TOUBRO LIMITED,LT,INR,44652051
+LTFH,L&T FINANCE HOLDINGS LTD,L&TFH,INR,92679222
+LUPIN,LUPIN LTD,LUPIN,INR,78429379
+MANAPPURA,MANAPPURAM FINANCE LTD,MANAPPURAM,INR,74960455
+MARICO,MARICO LIMITED,MARICO,INR,56987333
+MARUTI,MARUTI SUZUKI INDIA LTD,MARUTI,INR,44652108
+MCDOWELL-,UNITED SPIRITS LTD,MCDOWELL-N,INR,321318372
+MFSL,MAX FINANCIAL SERVICES LTD,MFSL,INR,56987139
+MGL,MAHANAGAR GAS LTD,MGL,INR,239032762
+MINDTREE,MINDTREE LIMITED,MINDTREE,INR,56987628
+MM,MAHINDRA & MAHINDRA LIMITED,M&M,INR,73894413
+MMFIN,MAHINDRA & MAHINDRA FIN SECS,M&MFIN,INR,122139383
+MOTHERSUM,MOTHERSON SUMI SYSTEMS LTD,MOTHERSUMI,INR,56987580
+MRF,MRF LIMITED,MRF,INR,56987338
+MUTHOOTFI,MUTHOOT FINANCE LTD,MUTHOOTFIN,INR,87790418
+NATIONALU,NATIONAL ALUMINIUM CO LTD,NATIONALUM,INR,85433002
+NAUKRI,INFO EDGE INDIA LTD,NAUKRI,INR,56985708
+NESTLEIND,NESTLE INDIA LIMITED,NESTLEIND,INR,64687316
+NMDC,NMDC LIMITED,NMDC,INR,56984653
+NTPC,NTPC LIMITED,NTPC,INR,49982069
+ONGC,OIL & NATURAL GAS CORP LTD,ONGC,INR,83845926
+PAGEIND,PAGE INDUSTRIES LTD,PAGEIND,INR,56986688
+PEL,PIRAMAL ENTERPRISES LTD,PEL,INR,56986920
+PETRONET,PETRONET LNG LTD,PETRONET,INR,56986960
+PFC,POWER FINANCE CORPORATION,PFC,INR,56986985
+PIDILITIN,PIDILITE INDUSTRIES LTD,PIDILITIND,INR,56986764
+PNB,PUNJAB NATIONAL BANK,PNB,INR,176235422
+POWERGRID,POWER GRID CORP. OF INDIA LTD.,POWERGRID,INR,50486118
+PVR,PVR LTD,PVR,INR,56987108
+RAMCOCEM,RAMCO CEMENTS LTD/THE,RAMCOCEM,INR,56987148
+RBLBANK,RBL BANK LTD,RBLBANK,INR,246723708
+RECLTD,REC LTD,RECLTD,INR,56984788
+RELIANCE,RELIANCE INDUSTRIES LIMITED,RELIANCE,INR,44652000
+SAIL,STEEL AUTHORITY OF INDIA,SAIL,INR,44652045
+SBILIFE,SBI LIFE INSURANCE CO LTD,SBILIFE,INR,291082784
+SBIN,STATE BANK OF INDIA,SBIN,INR,173418084
+SHREECEM,SHREE CEMENT LIMITED,SHREECEM,INR,56989171
+SIEMENS,SIEMENS LTD,SIEMENS,INR,44652105
+SRF,SRF LIMITED,SRF,INR,56989180
+SRTRANSFI,SHRIRAM TRANSPORT FINANCE,SRTRANSFIN,INR,56988999
+SUNPHARMA,SUN PHARMACEUTICAL INDUS,SUNPHARMA,INR,81351036
+SUNTV,SUN TV NETWORK LIMITED,SUNTV,INR,56989289
+TATACHEM,TATA CHEMICALS LIMITED,TATACHEM,INR,56986895
+TATACONSU,TATA CONSUMER PRODUCTS LTD,TATACONSUM,INR,76814483
+TATAMOTOR,TATA MOTORS LTD,TATAMOTORS,INR,94127580
+TATAPOWER,TATA POWER CO LTD,TATAPOWER,INR,95168502
+TATASTEEL,TATA STEEL LTD,TATASTEEL,INR,44652069
+TCS,TATA CONSULTANCY SVCS LTD,TCS,INR,44652015
+TECHM,TECH MAHINDRA LTD,TECHM,INR,187395599
+TITAN,TITAN CO LTD,TITAN,INR,90181242
+TORNTPHAR,TORRENT PHARMACEUTICALS LTD,TORNTPHARM,INR,56986878
+TORNTPOWE,TORRENT POWER LTD,TORNTPOWER,INR,56986853
+TVSMOTOR,TVS MOTOR CO LTD,TVSMOTOR,INR,56986838
+UBL,UNITED BREWERIES LTD,UBL,INR,56983324
+ULTRACEMC,ULTRATECH CEMENT LTD,ULTRACEMCO,INR,56983383
+UPL,UPL LTD,UPL,INR,56983369
+VEDL,VEDANTA LTD,VEDL,INR,56988951
+VOLTAS,VOLTAS LIMITED,VOLTAS,INR,56983785
+WIPRO,WIPRO LTD,WIPRO,INR,44652030
+ZEEL,ZEE ENTERTAINMENT ENTERPRISE,ZEEL,INR,44652165
